Share platform runtime env builder and run polling across executors
Hoist buildPlatformRuntimeEnv into scheduled-job-route.ts and extract
waitForRunCompletion into a new run-polling.ts so scheduler and
automation executors stop carrying byte-identical private copies.
